Avoid dynamic dispatch in inner loop using trait-based dispatch.
As the impl Fn(...) -> ... will resolve to fn(...) -> ..., the estimator and distance
functions will be called via function pointers, i.e. via dynamic dispatch.

To avoid this, two traits are defined and the various algorithms are parametrized
on those traits using type-level parameters. This way, the actual implementations
will be monomorphized for the relevant combinations of estimator and distance function.