Initial kriging commit
This is the first step toward implementing the full kriging workflow in
one of the GSTools core packages (which is the Rust one in this case,
obviously). For now, only simple kriging on isotropic fields is
implemented. This commit also includes a few early pieces for the
CovModel implemented in the core package.